action: use pkgbase rather than pkgname for deps
Previously, we assumed that the name of directory (the pkgbase) of
packages was the same as pkgbase. Unfortunately, this assumption is not
correct and has been causing some failures (such as with inkscape-git,
which depends on libmagick6, whose pkgbase is actually imagemagick6).

Instead, at the beginning of each build, we now walk through every
package in the repository and build a mapping of pkgname to pkgbase,
stored as a bash associative array. This mapping is then used when
accessing directories for a particular package.

Note that this now requires that .SRCINFO be kept up-to-date (or at
least semi-up-to-date), but this should not be a problem because the AUR
already enforces this.