autosign: config option to skip attestation check
Although doing the attestation check is nice here as it provides
additional security in the event of partial compromise of GitHub, it is
not strictly necessary. It currently relies on calling out to an
external dependency, the gh binary, which is not ideal for running in a
stripped down environment. It also creates additional attack surface on
the autosign side.

When this option is enabled, this tool can be run without any external
dependencies. When go-github gets support for doing this directly, we
can reevaluate whether or not we want to keep this option.

It is of course possible to run the attestation check at other points in
the process, such as before adding packages to the repository.