Add listen mode, unify error messages
This mode currently only supports working with webhook events proxied
through smee.io. On the GitHub side, a webhook is configured to post
workflow_run events (only) to the appropriate endpoint. We validate the
signature, check that the action is completed, and if so, process the
artifacts generated in that workflow run.