autosign: implement native Go attestation check
Recently, go-github added the ability to list attestations associated
with a GitHub repository. By combining this with sigstore-go to perform
verification, we can now verify the attestation of packages without
having to rely on the GitHub CLI.

For now, we ship a static trusted_root.json that is embedded into the
binary. In the future, we may want to make this configurable.