Remove old controller stuff
This didn't make a ton of sense with the way I'm using this these days
and needs refactoring anyway. It's perfectly fine to just run ssh
multiple times on the caller end. By modifying the code to background
the Docker commands, this can be done in an asynchronous manner.

The changes to build_pkgtree.py have not yet been tested. It would
probably also make sense to pin the build image; this can be handled on
the client side as well.